refactor: Rename `…/event-detection/event-types/cloudwatchEvent.js`
