+++ sdk-js/src/lib/event-detection/event-types/cloudwatch-event.js
