+++ sdk-js/src/lib/event-detection/index.js
@@ -5,7 +5,7 @@ const apiGateway = require('./event-types/apiGateway');
 const apiGatewayV2 = require('./event-types/api-gateway-v2');
 const customAuthorizer = require('./event-types/customAuthorizer');
 const cloudFront = require('./event-types/cloudFront');
-const cloudwatchEvent = require('./event-types/cloudwatchEvent');
+const cloudwatchEvent = require('./event-types/cloudwatch-event');
 const cloudwatchLogs = require('./event-types/cloudwatchLog');
 const dynamodb = require('./event-types/dynamodb');
 const firehose = require('./event-types/firehose');
